Fix DeltaTo*Minutes fields reading two bytes instead of one
DeltaToCarInFrontMinutes and DeltaToRaceLeaderMinutes were read via
ReadUnaligned<ushort> for the 2024, 2025, and 2026 lap data packets,
but the offset was only advanced by one byte as required by the
uint8 spec for these fields. The extra byte pulled in the low byte of
the following field whenever it was non-zero, producing wrong values.
Both fields are now read via ReadUnaligned<byte>, matching the single
byte the game actually sends and the existing offset advance.

+++ F1Server.Core/Packets/PacketToObject/PacketToLapData.cs
@@ -857,7 +857,7 @@ internal class PacketToLapData(PacketHeader packetHeader) : PacketToXBase(packet
 
             actOffset += ConstData.TypeUInt16;
 
-            lapData.DeltaToCarInFrontMinutes = Unsafe.ReadUnaligned<ushort>(ref Unsafe.Add(ref dataPacket, actOffset));
+            lapData.DeltaToCarInFrontMinutes = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
             actOffset += ConstData.TypeUInt8;
 
@@ -865,7 +865,7 @@ internal class PacketToLapData(PacketHeader packetHeader) : PacketToXBase(packet
 
             actOffset += ConstData.TypeUInt16;
 
-            lapData.DeltaToRaceLeaderMinutes = Unsafe.ReadUnaligned<ushort>(ref Unsafe.Add(ref dataPacket, actOffset));
+            lapData.DeltaToRaceLeaderMinutes = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
             actOffset += ConstData.TypeUInt8;
 
@@ -1014,7 +1014,7 @@ internal class PacketToLapData(PacketHeader packetHeader) : PacketToXBase(packet
 
             actOffset += ConstData.TypeUInt16;
 
-            lapData.DeltaToCarInFrontMinutes = Unsafe.ReadUnaligned<ushort>(ref Unsafe.Add(ref dataPacket, actOffset));
+            lapData.DeltaToCarInFrontMinutes = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
             actOffset += ConstData.TypeUInt8;
 
@@ -1022,7 +1022,7 @@ internal class PacketToLapData(PacketHeader packetHeader) : PacketToXBase(packet
 
             actOffset += ConstData.TypeUInt16;
 
-            lapData.DeltaToRaceLeaderMinutes = Unsafe.ReadUnaligned<ushort>(ref Unsafe.Add(ref dataPacket, actOffset));
+            lapData.DeltaToRaceLeaderMinutes = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
             actOffset += ConstData.TypeUInt8;
 
@@ -1171,7 +1171,7 @@ internal class PacketToLapData(PacketHeader packetHeader) : PacketToXBase(packet
 
             actOffset += ConstData.TypeUInt16;
 
-            lapData.DeltaToCarInFrontMinutes = Unsafe.ReadUnaligned<ushort>(ref Unsafe.Add(ref dataPacket, actOffset));
+            lapData.DeltaToCarInFrontMinutes = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
             actOffset += ConstData.TypeUInt8;
 
@@ -1179,7 +1179,7 @@ internal class PacketToLapData(PacketHeader packetHeader) : PacketToXBase(packet
 
             actOffset += ConstData.TypeUInt16;
 
-            lapData.DeltaToRaceLeaderMinutes = Unsafe.ReadUnaligned<ushort>(ref Unsafe.Add(ref dataPacket, actOffset));
+            lapData.DeltaToRaceLeaderMinutes = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
             actOffset += ConstData.TypeUInt8;
